(bzr r3627.2.18)

+++ plugins/compiztoolbox/include/compiztoolbox/compiztoolbox.h
@@ -162,6 +162,6 @@ class BaseSwitchWindow
 };
 
 extern const unsigned short ICON_SIZE;
-extern const unsigned int MAX_ICON_SIZE;
+extern const unsigned short MAX_ICON_SIZE;
 
 #endif
